refactor dialLoop into fetchMsg
Move all logic into a non-looping method that dials the server, fetches
a binary blob, unmarshals it into a message and returns that message.

It logs all errors, using the returned errors to indicate retry, the context
being cancelled or nil for a successful call.

This allowed the test to be refactored and made much simpler.